Charge tax to towns, add tax to nation status screen.
