+++ src/main/java/com/gmail/goosius/siegewar/listeners/SiegeWarStatusScreenListener.java
@@ -61,6 +61,11 @@ public class SiegeWarStatusScreenListener implements Listener {
 		if (SiegeWarSettings.getWarSiegeEnabled()) {
 			final Translator translator = Translator.locale(Translation.getLocale(event.getCommandSender()));
 			Nation nation = event.getNation();
+			double occupationTax = NationMetaDataController.getNationOccupationTax(nation); 
+			if (occupationTax > 0) {
+				Component comp = Component.text(translator.of("status_nation_occupation_tax", occupationTax)).appendNewline();
+				event.getStatusScreen().addComponentOf("siegeWarOccupationTax", comp);
+			}
 			List<String> out = new ArrayList<>();
 			// Occupied Home Towns[3]: Town1, Town2, Town3
 			List<Town> occupiedHomeTowns = TownOccupationController.getOccupiedHomeTowns(nation);
