+++ src/main/java/com/gmail/goosius/siegewar/listeners/SiegeWarTownyEventListener.java
@@ -119,6 +119,7 @@ public class SiegeWarTownyEventListener implements Listener {
 			if (SiegeWarSettings.isPlunderPaidOutOverDays()) {
 				SiegeWarMoneyUtil.payDailyPlunderDebt();
 			}
+			TownOccupationController.chargeNationOccupationTax();
 		}
 	}
